1_Google_Gemini_LLM_TextToSQL/requirements.txt
python-dotenv==1.0.0
streamlit==1.22.0
pandas==2.2.2
google-generativeai==0.6.0
pyodbc==5.1.0